+++ include/xieite/geometry/getBoundingBox.hpp
@@ -11,8 +11,8 @@
 namespace xieite::geometry {
 	[[nodiscard]]
 	constexpr xieite::geometry::Polygon getBoundingBox(const xieite::geometry::Ellipse& ellipse) noexcept {
-		const long double a = xieite::math::squareRoot(ellipse.radii.x * ellipse.radii.x * std::cos(ellipse.rotation) * std::cos(ellipse.rotation) + ellipse.radii.y * ellipse.radii.y * std::cos(ellipse.rotation + xieite::math::pi<long double> / 2) * std::cos(ellipse.rotation + xieite::math::pi<long double> / 2));
-		const long double b = xieite::math::squareRoot(ellipse.radii.x * ellipse.radii.x * std::sin(ellipse.rotation) * std::sin(ellipse.rotation) + ellipse.radii.y * ellipse.radii.y * std::sin(ellipse.rotation + xieite::math::pi<long double> / 2) * std::sin(ellipse.rotation + xieite::math::pi<long double> / 2));
+		const double a = xieite::math::squareRoot(ellipse.radii.x * ellipse.radii.x * std::cos(ellipse.rotation) * std::cos(ellipse.rotation) + ellipse.radii.y * ellipse.radii.y * std::cos(ellipse.rotation + xieite::math::pi<double> / 2) * std::cos(ellipse.rotation + xieite::math::pi<double> / 2));
+		const double b = xieite::math::squareRoot(ellipse.radii.x * ellipse.radii.x * std::sin(ellipse.rotation) * std::sin(ellipse.rotation) + ellipse.radii.y * ellipse.radii.y * std::sin(ellipse.rotation + xieite::math::pi<double> / 2) * std::sin(ellipse.rotation + xieite::math::pi<double> / 2));
 		return xieite::geometry::Polygon(std::vector<xieite::geometry::Point> {
 			xieite::geometry::Point(ellipse.center.x - a, ellipse.center.y - b),
 			xieite::geometry::Point(ellipse.center.x + a, ellipse.center.y - b),
